backend3/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
bcrypt==3.2.2
beautifulsoup4==4.14.3
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cobble==0.1.4
cryptography==46.0.5
cuda-bindings==12.9.4
cuda-pathfinder==1.3.3
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
faiss-cpu==1.13.2
fastapi==0.128.7
filelock==3.20.3
filetype==1.2.0
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.2.0
google-auth==2.48.0
google-genai==1.62.0
greenlet==3.3.1
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==0.36.2
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.10
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.10
langchain-google-genai==4.2.0
langchain-huggingface==1.2.0
langchain-text-splitters==1.1.0
langgraph==1.0.8
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.5
langsmith==0.7.1
lxml==6.0.2
mammoth==1.11.0
MarkupSafe==3.0.3
marshmallow==3.26.2
motor==3.7.1
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
networkx==3.6.1
numpy==2.4.2
onnxruntime==1.24.1
opencv-python==4.13.0.92
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
passlib==1.7.4
pillow==12.1.0
propcache==0.4.1
protobuf==6.33.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pyclipper==1.4.0
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pymongo==4.16.0
PyMuPDF==1.26.7
python-docx==1.2.0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
PyYAML==6.0.3
rapidocr-onnxruntime==1.4.4
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
sentence-transformers==5.2.2
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.46
starlette==0.52.1
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.3
transformers==4.57.6
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.0
uvicorn==0.40.0
watchfiles==1.1.1
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0

pandas
openpyxl
pypdf
langgraph-checkpoint-sqlite
tavily-python
# aiohappyeyeballs==2.6.1
# aiohttp==3.13.3
# aiosignal==1.4.0
# annotated-doc==0.0.4
# annotated-types==0.7.0
# anyio==4.12.1
# attrs==25.4.0
# bcrypt==3.2.2
# beautifulsoup4==4.14.3
# certifi==2026.1.4
# cffi==2.0.0
# charset-normalizer==3.4.4
# click==8.3.1
# cobble==0.1.4
# cryptography==46.0.5
# cuda-bindings==12.9.4
# cuda-pathfinder==1.3.3
# dataclasses-json==0.6.7
# distro==1.9.0
# dnspython==2.8.0
# ecdsa==0.19.1
# email-validator==2.3.0
# faiss-cpu==1.13.2
# fastapi==0.128.7
# filelock==3.20.3
# filetype==1.2.0
# flatbuffers==25.12.19
# frozenlist==1.8.0
# fsspec==2026.2.0
# google-auth==2.48.0
# google-genai==1.62.0
# greenlet==3.3.1
# h11==0.16.0
# hf-xet==1.2.0
# httpcore==1.0.9
# httptools==0.7.1
# httpx==0.28.1
# httpx-sse==0.4.3
# huggingface_hub==0.36.2
# idna==3.11
# Jinja2==3.1.6
# joblib==1.5.3
# jsonpatch==1.33
# jsonpointer==3.0.0
# langchain==1.2.10
# langchain-classic==1.0.1
# langchain-community==0.4.1
# langchain-core==1.2.10
# langchain-google-genai==4.2.0
# langchain-huggingface==1.2.0
# langchain-text-splitters==1.1.0
# langgraph==1.0.8
# langgraph-checkpoint==4.0.0
# langgraph-prebuilt==1.0.7
# langgraph-sdk==0.3.5
# langsmith==0.7.1
# lxml==6.0.2
# mammoth==1.11.0
# MarkupSafe==3.0.3
# marshmallow==3.26.2
# motor==3.7.1
# mpmath==1.3.0
# multidict==6.7.1
# mypy_extensions==1.1.0
# networkx==3.6.1
# numpy==2.4.2
# nvidia-cublas-cu12==12.8.4.1
# nvidia-cuda-cupti-cu12==12.8.90
# nvidia-cuda-nvrtc-cu12==12.8.93
# nvidia-cuda-runtime-cu12==12.8.90
# nvidia-cudnn-cu12==9.10.2.21
# nvidia-cufft-cu12==11.3.3.83
# nvidia-cufile-cu12==1.13.1.3
# nvidia-curand-cu12==10.3.9.90
# nvidia-cusolver-cu12==11.7.3.90
# nvidia-cusparse-cu12==12.5.8.93
# nvidia-cusparselt-cu12==0.7.1
# nvidia-nccl-cu12==2.27.5
# nvidia-nvjitlink-cu12==12.8.93
# nvidia-nvshmem-cu12==3.4.5
# nvidia-nvtx-cu12==12.8.90
# onnxruntime==1.24.1
# opencv-python==4.13.0.92
# orjson==3.11.7
# ormsgpack==1.12.2
# packaging==26.0
# passlib==1.7.4
# pillow==12.1.0
# propcache==0.4.1
# protobuf==6.33.5
# pyasn1==0.6.2
# pyasn1_modules==0.4.2
# pyclipper==1.4.0
# pycparser==3.0
# pydantic==2.12.5
# pydantic-settings==2.12.0
# pydantic_core==2.41.5
# pymongo==4.16.0
# PyMuPDF==1.26.7
# python-docx==1.2.0
# python-dotenv==1.2.1
# python-jose==3.5.0
# python-multipart==0.0.22
# PyYAML==6.0.3
# rapidocr-onnxruntime==1.4.4
# regex==2026.1.15
# requests==2.32.5
# requests-toolbelt==1.0.0
# rsa==4.9.1
# safetensors==0.7.0
# scikit-learn==1.8.0
# scipy==1.17.0
# sentence-transformers==5.2.2
# shapely==2.1.2
# six==1.17.0
# sniffio==1.3.1
# soupsieve==2.8.3
# SQLAlchemy==2.0.46
# starlette==0.52.1
# sympy==1.14.0
# tenacity==9.1.4
# threadpoolctl==3.6.0
# tokenizers==0.22.2
# torch==2.10.0
# tqdm==4.67.3
# transformers==4.57.6
# triton==3.6.0
# typing-inspect==0.9.0
# typing-inspection==0.4.2
# typing_extensions==4.15.0
# urllib3==2.6.3
# uuid_utils==0.14.0
# uvicorn==0.40.0
# uvloop==0.22.1
# watchfiles==1.1.1
# websockets==15.0.1
# xxhash==3.6.0
# yarl==1.22.0
# zstandard==0.25.0

# pandas
# openpyxl
# pypdf
# langgraph-checkpoint-sqlite
# tavily-python
